refactor: move websocket to a different file
